fix dtype inference: single source of truth from model weights
HouseholderOrthogonal.get_matrix() was creating torch.eye() without dtype,
defaulting to float32. When model weights are bf16, this caused dtype
mismatch in matmul operations. Fix: inherit dtype from self.vs parameter.

+++ src/model.py
@@ -240,7 +240,7 @@ class HouseholderOrthogonal(nn.Module):
 
     def get_matrix(self):
         # Start with Identity
-        Q = torch.eye(self.dim, device=self.vs.device)
+        Q = torch.eye(self.dim, device=self.vs.device, dtype=self.vs.dtype)
         # Iteratively apply reflections: H = I - 2vv^T / ||v||^2
         for i in range(self.vs.shape[0]):
             v = self.vs[i].unsqueeze(1)
